backend-flask/requirements.txt
Flask==3.0.3
Pillow==10.3.0
python-dotenv==1.0.1
Werkzeug==3.0.3
